major reorg of files
no new functionality here - just splitting up the one large file into 8-10 smaller files.  Now have a models directory and one file per model, as well as a tables directory and one file per table.  Remember this is only a convention, so mixing in other aspects of the simulation framework where appropriate is expected.
