+++ activitysim/defaults/__init__.py
@@ -1,0 +1,7 @@
+# ActivitySim
+# Copyright (C) 2014-2015 Synthicity, LLC
+# See full license in LICENSE.txt.
+
+import misc
+import tables
+import models
