+++ activitysim/defaults/models/scheduling.py
@@ -1,3 +1,8 @@
+import os
+import pandas as pd
+import urbansim.sim.simulation as sim
+from activitysim import activitysim as asim
+
 """
 This model predicts the departure time and duration of each activity for
 mandatory tours
