+++ activitysim/defaults/models/workplace_location.py
@@ -1,3 +1,9 @@
+import os
+import urbansim.sim.simulation as sim
+from activitysim import activitysim as asim
+import pandas as pd
+import numpy as np
+
 """
 The workplace location model predicts the zones in which various people will
 work.  Interestingly there's not really any supply side to this model - we
