+++ activitysim/defaults/tables/__init__.py
@@ -1,0 +1,9 @@
+# ActivitySim
+# Copyright (C) 2014-2015 Synthicity, LLC
+# See full license in LICENSE.txt.
+
+import households
+import persons
+import landuse
+import skims
+import accessibility
