+++ activitysim/defaults/tables/persons.py
@@ -1,3 +1,9 @@
+import pandas as pd
+import numpy as np
+import urbansim.sim.simulation as sim
+import urbansim.utils.misc as usim_misc
+
+
 @sim.table(cache=True)
 def persons(store, settings, households):
     df = store["persons"]
