+++ activitysim/defaults/tables/skims.py
@@ -1,3 +1,8 @@
+import urbansim.sim.simulation as sim
+import openmatrix as omx
+from activitysim import skim
+import os
+
 """
 Read in the omx files and create the skim objects
 """
